Update patch set 1
Patch Set 1:

This was disabled on victara by crpalmer months before it was on 8226-common, and I'm sure many others disabled it before then. Disabling a feature that doesn't work is a fairly obvious thing to do (it was the first thing I tried without thinking or looking anywhere else).

Patch-set: 1