Update patch set 1
Patch Set 1: Code-Review+2 Verified+1

The open source charge_only_mode is archaic and nobody had bothered to update it for devices from this decade. This change allows it to properly detect power uevents (such as unplugging the charger) so that it can do what it needs to (eg. turn off when charger is unplugged). There are a lot more updates and improvements that can be made to the charge-only mode, but at least it can work properly now with this change, and not stay awake draining the battery when you unplug the charger.

Patch-set: 1
Reviewer: Sultan Qasim Khan <13881@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=+1
Label: Code-Review=+2